Tests results
It seems that putting about 1,2,5, 10k recruits  don't influence speed of working of
site on Heroku from 1MB ADSL connection noticably (that's what I used)
for guests (I didn't manage to get JMeter to log in and perform some
actions as recruit).

As recruiter I couldn't view homepage (because when there were ~2k users
it tried to show them all which took >30s to render which made Heroku
stop it).